Provide coverage threshold

+++ assembler/src/debruijn/genomic_info.hpp
@@ -20,8 +20,8 @@ class GenomicInfo {
     double ec_bound() const { return ec_bound_; }
     void set_ec_bound(double ec_bound) { ec_bound_ = ec_bound; }
 
-    size_t trusted_bound() const { return trusted_bound_; }
-    void set_trusted_bound(size_t trusted_bound) { trusted_bound_ = trusted_bound; }
+    double trusted_bound() const { return trusted_bound_; }
+    void set_trusted_bound(double trusted_bound) { trusted_bound_ = trusted_bound; }
 
     bool Load(const std::string &filename);
     void Save(const std::string &filename) const;
@@ -31,7 +31,7 @@ class GenomicInfo {
     size_t genome_size_;
     double estimated_mean_;
     double ec_bound_;
-    size_t trusted_bound_;
+    double trusted_bound_;
 };
 
 #endif
